requirements.txt
pyTelegramBotAPI==3.2.1
PyYAML==3.12
emoji==0.4.5
